+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "suspension"
+version = "0.1.0"
+requires-python = ">=3.8"
+
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["suspension"]
