+++ scripts/main.py
@@ -1,8 +1,8 @@
 import argparse
 import time
 
-from utils.bump_generator import generate_elliptical_cylinder_stl
-from utils.update_mjcf import insert_stl_to_mjcf
+from suspension.utils.bump_generator import generate_elliptical_cylinder_stl
+from suspension.utils.update_mjcf import insert_stl_to_mjcf
 import mujoco
 import mujoco.viewer
